Class cleanup and file restructure
