+++ classes/shapes.js
@@ -1,0 +1,10 @@
+class Shape {
+    //Parent class
+    constructor(logoLetters, logoLettersColor,logoColor) 
+    {
+        this.logoLetters = logoLetters;
+        this.logoLettersColor = logoLettersColor;
+        this.logoColor = logoColor;
+    }}
+
+
